Remove wrong widgets
