+++ lib/creatego_theme.dart
@@ -18,15 +18,11 @@ export 'src/widgets/yoshop_sys_components/spaced_column/spaced_column.dart';
 export 'src/widgets/yoshop_sys_components/spaced_row/spaced_row.dart';
 
 //Yoshop Sys
-export 'widgets/yoshop_sys_components/sidebar/ys_sidebar.dart';
-export 'widgets/yoshop_sys_components/buttons/ys_button.dart';
-export 'widgets/yoshop_sys_components/toogles/ys_toogle.dart';
-export 'widgets/yoshop_sys_components/checkboxes/ys_checkbox.dart';
-export 'widgets/yoshop_sys_components/datatables/default_datatable.dart';
-export 'widgets/yoshop_sys_components/dropdowns/ys_dropdown.dart';
-export 'widgets/yoshop_sys_components/inputfields/ys_inputfield.dart';
-export 'widgets/yoshop_sys_components/status/ys_status.dart';
-
+export 'src/widgets/yoshop_sys_components/sidebar/ys_sidebar.dart';
+export 'src/widgets/yoshop_sys_components/buttons/ys_button.dart';
+export 'src/widgets/yoshop_sys_components/toogles/ys_toogle.dart';
+export 'src/widgets/yoshop_sys_components/checkboxes/ys_checkbox.dart';
+export 'src/widgets/yoshop_sys_components/datatables/default_datatable.dart';
 
 //Finenex
 export 'src/widgets/fx_components/fx_box_content.dart';
