--- lib/widgets/yoshop_sys_components/spaced_row/spaced_row.dart
@@ -1,35 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-
-class SpacedRow extends StatelessWidget {
-  //Do not add screenUtil, Just pass double value
-  double? horizontalSpace;
-  List<Widget> children;
-  MainAxisAlignment? mainAxisAlignment;
-  CrossAxisAlignment? crossAxisAlignment;
-
-  SpacedRow(
-      {this.horizontalSpace = 0.0,
-      required this.children,
-      this.mainAxisAlignment = MainAxisAlignment.start,
-      this.crossAxisAlignment = CrossAxisAlignment.start});
-
-  @override
-  Widget build(BuildContext context) {
-    List<Widget> widgets = [];
-    for (var element in children) {
-      widgets.add(element);
-      if (children.last == element) {
-      } else {
-        widgets.add(SizedBox(
-          width: horizontalSpace!,
-        ));
-      }
-    }
-    return Row(
-      mainAxisAlignment: mainAxisAlignment!,
-      crossAxisAlignment: crossAxisAlignment!,
-      children: widgets,
-    );
-  }
-}
